pyproject.toml
[tool.poetry]
name = "autograder-pipeline-fe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
langchain = "^0.3.4"
azure-identity = "^1.19.0"
langchain-openai = "^0.2.3"
python-dotenv = "^1.0.1"
langchain-groq = "^0.2.0"
pymongo = "^4.10.1"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
